Add Resque/ActiveJob support
Our old Resque plugin doesn't work with ActiveJob,
I've added a new plugin to support Resque in an ActiveJob scenario.

It uses the ActiveJob `around_perform` callback
to wrap it in an `Appsignal.monitor_single_transaction` block,
making sure the extension pushes the data to the agent before it closes.