Add AppSignal powered params filtering
This adds support to filter params in non-Rails applications, as we rely
on the built-in Rails params filtering then.

This params filtering supports key based filtering for hashes, the
values of which will be replaced with the "[FILTERED]" value. There's
support for nested hashes and nested hashes in arrays. Basically, any
params hash will be filtered.

To use this filtering, add the following to your `config/appsignal.yml`
file in the environment group you want it to apply. The
`filter_parameters` value is an Array of strings.

```yml
production:
  filter_parameters:
    - password
    - confirm_password
```