Use GC::Profiler to track event and transaction GC
Instead of installing the Garbage Collection event hooks, setting
enable_gc_instrumentation in the AppSignal config now enables
GC::Profiler.

Appsignal::GarbageCollectionProfiler takes data from GC::Profiler
to report it to the extension whenever a transaction or event is started
or finished. It will also make sure to clear the internal garbage
collection profiler to stop it from leaking.

When the internal garbage collection profiler is slow to clear the total
GC time, Appsignal::GarbageCollectionProfiler#total_time uses a Mutex to
lock and make sure GC time is not counted twice.

The test for this sets a clear delay, which is used by FakeGCProfiler to
pause between returning the total time and clearing it. The test sets a
total time for the first thread, which gets cleared and should be 0 for
the second one.

In the tests, GC::Profiler.enable is stubbed in with gc instrumentation
enabled to make sure the test suite doesn't touch GC::Profiler.

FakeGCProfiler behaves like Ruby's internal GC::Profiler, but has a
writable @total_time for testing purposes. Using this instead of
stubbing the results of GC::Profiler.total_time allows for more a more
accurate replication of GC::Profiler's behavior by implementing a
working .clear method.