Ensure capistrano2 and 3 have configuration parity and add better error logging
The documentation states that you can set a `stage` config and appsignal will treat the deploy as that environment, however capistrano2 did not have support for this feature.

This commit adds support for the `stage` variable.

It also improves error logging by adding environment name to the message when the environment is not active.