Allow config of ca_file_path
Needed to be configured in the agent as well so add a config option for
it to be set to the environment.

Test if ca_file_path file exists and is readable
If not, ignore the settings because if set with a faulty value it would
crash the process.
Log a warning and test the output of the log to see if it ignores it.