Don't fallback on /tmp dir if it's not writable
Currently if the set log path, or the project path, are not writable we
always fallback on the `/tmp/appsignal.log` fallback path. This breaks
if the fallback path is also not writable.

This change adds a permission check on the /tmp fallback path and
returns `nil` if no suitable log path can be found.

Update test to handle more cases.