Change load order on config
The new load order will be:

1. System detected settings (such as :running_in_container)
2. Defaults (see Config::DEFAULTS)
3. Initial config given to Config initializer
4. appsignal.yml config file
5. Environment variables

The idea behind this change is:
1. we thought it was already working this way.. and
2. this allows the highest user configurable method to always be
leading; the environment variables. A config file is harder to change
per system while a config variable is easier to change per system.

This change is a little bigger than expected because I began updating
the spec, found duplicate specs, strangely nested specs and not a good
story of what overrides what.

That's now changed.

I updated the specs to reflect this change, remove duplicates, fix the
nesting, remove nesting where it wasn't necessary and reorder specs to
reflect the hierarchy of config loading. Also updated the specs to use
the same `expect` syntax and don't use the `its` syntax in preparation
of an upgrade to RSpec 3.