Refactor rake task to not stub so much
- Do not manually install the Rake task, but use `.load_hooks` instead.
  This prevents hooks from being installed multiple times.
- Do not stub the config
- Do not stub `Appsignal.active?`
- Do not stub the Rake application and config
- Do not stub the Rake task, but enhance it with the expected behavior
- Do not stub calls to the transaction: `set_action`, `set_error`,
`complete`
- Do not stub the transaction id
- Test if the error is reraised after being set on the transaction
- Test if complete and stop happen in the right order

There's some unexpected behavior in this test as well, and possibly many
others. `background_job_transaction` creates a transaction, but sets it
as the current one as well. `create` is being used instead of `new`.
This is the reason this would have worked at all:
https://github.com/appsignal/appsignal-ruby/blob/ac6aa73a13903915331e6c27d596a705b3d804d0/spec/lib/appsignal/hooks/rake_spec.rb#L44-L54
In this test a transaction creation call is being expected and messages
are being expected on another one. Thus the specs don't properly test if
the transactions are being created since one already exists.