notify_of_deploy name option is leading when given
When notify_of_deploy was run in a path with a valid config or with the
environment variable APPSIGNAL_APP_NAME these two values would be
leading. The name option would only be applied when the config and the
environment had no mention of it.

Since --name is the last thing a user would set when running this
command, the config file already exists and the environment is set on
the machine, it should override any previous config to be applied.

This means setting it on the initial_config of the config wouldn't do
anything in most cases.