Refactor notify_of_deploy spec
Run command through full CLI stack on test, testing the whole process
removing the need for a separate cli_spec.

Environment is required only when its not set through the environment,
APPSIGNAL_APP_ENV.

Name is required if the config (either from the config file or the
environment) does not include a name.

Config is not passed to the CLI but initialized in the notify_of_deploy
command itself allowing for better control over how its initialized per
command. Only updating notify_of_deploy for now to reduce the scope of
the change.