Use consistent log formatter per log type. PR #203
The custom log format for the stdout logger was used primary for Heroku.
On Heroku the logs are prefixed with a timestamp and this custom format
did not include one.

Not a lot of other libraries offer special log formats for Heroku and if
this is the only exception there is not a huge incentive to omit the
timestamp from the log.