Fix APPSIGNAL_PUSH_API_KEY activation behavior
Regression fix compared to version 1.3.6 and 2.0.0.

This fix will restore the original behavior of APPSIGNAL_PUSH_API_KEY
setting Appsignal.active to true if it's in the environment, but it will
be unset again if the appsignal.yml configuration file is loaded as
well and sets `active` to another value.

By moving the APPSIGNAL_PUSH_API_KEY "activation" behavior from the last
step (loading environment variables) to the second step (system detected
settings) the original behavior from gem =< 1.3.6 is restored.

If the behavior is located in the last step any other configuration step
setting the `active` config would be undone.