Fix log path for Rails apps in diagnose output
When an app is a Rails app, we now prefix the "log/" dir to the
log_path on Appsignal start.

To do so we need to initialize the config in the diagnose command rather
than calling `Appsignal.start` and let it automatically find the config.
This causes the `extension_loaded` check to be performed manually rather
than relying on `Appsignal.start` not setting a config if the extension
is not loaded.

Also added an extra path to the path checks for `log_dir_path`, the
parent directory of an AppSignal log file.