Write log path permission warnings to log
Previously it was written to STDOUT, but I think it's okay now to write
it to the log and avoid spamming the STDOUT with (not always) very
important warnings.

Especially useful for users now that we change the log path to `log/`
in the project's `root_path`. If people upgrade and they don't have a
`log/` path, because they are not using Rails and have not customized
their `log_path`, they won't see any warnings in STDOUT, but the log
file instead.

Notes:

- logger.warn returns true/false
- Moved the `.logger`, `.start_logger` and `.log_formatter` specs in
  `appsignal_spec.rb` from the context "with config and started" to the
  root level of the spec file, because they set their own config if
  needed and don't require a started Appsignal.