Setup Rubocop for code styleguide
This commit sets up the generated config with our own configuration
options. The violations to the styleguide are registered in
`rubocop_todo.yml`. These can be fixed on by one or in groups of related
rules. Some rules affect the output of others when not all are enabled.

This first commit also fixes the `Style/ConditionalAssignment` rule.