Fix request/env being nil/no-method in transaction
When a Transaction is created and the request object is nil, or it
doesn't implement the `#env` method, such as a Hash, this causes
`NoMethodError`s to be raised.

This change adds more defensive checks so that the transaction doesn't
crash when such an object is given.

This change also documents a lot of the methods involved so it's easier
for me and future developers to see what's going on. Documentation
format is YARD http://yardoc.org/