"Support" bigints by converting them into strings
We convert big integers to Ruby because C doesn't support integers
bigger than int64. We prefix the value with `bigint:` so that we know
the original type of the value and can display it in its original form.

Unable to do this conversion on the C-side I opted to do this conversion
on the Ruby side in the `Appsignal::Utils::Data` class.