Update transaction garbage collection spec
Move garbage collection profiler spec to
class methods describe block.

Update spec to ensure it does not leak to other tests,
by triggering a reset of `total_time`.

The leak was not visible previously because of the location of the
spec (just before #record_event spec) and because @ext.record_event
was not passing `self.class.garbage_collection_profiler.total_time`.

Once doing so, the #record_event spec would have to be written to
account for changes in `total_time`, hence the update and move.