Don't force install Redis hook at in spec
Calling the `.install` method directly always installs the hook even if
the install condition isn't `true`.

Let `Appsignal::Hooks.load_hooks` `try_to_install` the hook, which
checks if the `dependencies_present?` check is `true`.

Only run the "with redis" context when the Redis gem is present. This
worked before because the specs that ran before "without redis" unset
the "Redis" constant after the spec was ran even when it was originally
loaded from the Redis gem.