Stub JS error catching endpoint when not active
Related to issue https://github.com/appsignal/appsignal-ruby/issues/247

When AppSignal is not active, but does have frontend_error_catching
turned on, most applications will still send errors to this endpoint
because the JS environment doesn't know what env it is running in.

To prevent spamming the server logs with 404 errors in such a scenario,
stub the frontend error catching endpoint and return the 202 status code
with a message explaining it's not active.