Use better Padrino action names
The main fix in this change is the order in which the action name is
found. It now prefers the `Sinatra::Request#controller` and
`Sinatra::Request#action` rather than
`Sinatra::Request#route_obj#original_path` which is from older Padrino
versions.

Also refactor specs to use a real Padrino app and not just test the
middleware. This way we will find out sooner if we break the integration
with Padrino itself.