Transmitter now returns response object
Previously it only returned the status code of the response.
We'll need the response body as well for the diagnose endpoint to read
the support token from the response.

This introduces a change in the behavior of the Transmitter, but since
it's a private class and shouldn't be used by users directly it should
not be a problem.

Other classes that depend on the Transmitter are updated to rely on the
response object instead of the status code.