Extension returns agent diagnostics report
The extension returns a JSON object, as a string, which contains the
agent diagnostics report.

The report definition is defined in the diagnose CLI task so we can
check all the tests and display them in a human readable format.

This change makes it possible to return the agent diagnostics report
without capturing the STDOUT while the agent is running. This proved
difficult to do in Elixir, where we now call the agent executable
directly. That will be changed once this change is accepted.

Instead of printing the report to STDOUT the extension will now return
the report as a JSON object, which is formatted as a string which needs
to be parsed into JSON first.

This should no longer require `$stdout.reopen` in the tests to capture
all the CLI output. We'll leave that logic for now, as it captures the
output from the AppSignal extension as well, so it's useful in other
specs as well.

This report may be incomplete when the agent fails to start or something
else goes wrong. Any error will be captured and added to the agent
report.