No longer gzip request bodies send by transmitter
There's no more need to gzip everything we send to our servers. We use
it in so few requests it's hardly worth the effort. We only use the
transmitter for the Push API key validation, which has no body, and the
marker transmitter which is not used frequently and no longer the
recommended method to create markers.

The marker transmitter, used by `appsignal notify_of_deploy`, which is a
method of creating deploy markers which is no longer the recommended way
to do so. Instead use the `APP_REVISION` environment variable. See also:
http://docs.appsignal.com/application/markers/deploy-markers.html