Accept a default yes_or_no response
If we specify a default value for the yes_or_no question in the CLI and
no input is given (just an enter) we will accept the default value.

If it says: "yes or no (Y/n)" we accept "Y". This format does need to be
entered manually in the prompt.

This change also accepts users entering uppercase letters "Y" and "N",
and also the full "yes" and "no" words.