Fix ActiveSupport instrumentation
Some events are now not picked up when there's no subscriber for it.
This is true when there are no default Rails subscribers, such as in
projects without Rails.

Previously we had a subscriber for all events, except private events
prefixed with a exclamation mark `!`. This subscriber would make sure
all events would be picked up.

With the current setup though we don't capture all events if there's no
subscriber. The check in `ActiveSupport::Notifications.instrument`
prevents it from hitting
`ActiveSupport::Notifications::Instrumenter.instrument`, so we override
the method to ignore the check.