When JS endpoint payload is empty return 400 code
Problem originally reported in #261, but the exact error was not
reproduced. Regardless, there is a scenario where if the payload is
empty it crashes in the endpoint. Instead, rescue the error and return a
400 "bad request" response.

This should create a somewhat more pleasant development experience with
more descriptive error messages instead of cryptic errors such as:

```
$ JSON.parse("")
JSON::ParserError: A JSON text must at least contain two octets!
```