Track ActiveSupport events inside AS.instrument
Change the way our ActiveSupport::Notifications.instrument method
override picks up the events. Now we can start a transaction inside an
event and have the event be tracked afterwards, inside the transaction.
Needed to start a transaction in the ActionCable event before tracking
it, and adding it to a transaction.