Add ActionCable support
Support ActionCable by overriding the `#perform_action` method and
wrapping it in an AppSignal transaction. The `#perform_action` is called
by ActionCable once a message is send to the channel.

We also track subscribe and unsubscribe events as they can also contain
complex logic added by the user, which is something we want to track. To
track this two "around" callbacks are added to ActionCable on load.