Update Shoryuken integration
Fix the Shoryuken integration. Looks like the initial version was based
on our Sidekiq integration, but it does not work like that.

Updated to work against version 1 through 3 with Shoryuken and this was
the behavior I could determine from the library.

The original request that brought the problem to our attention was to
support string (or any other type than a Hash) as a body for the worker.
This is now supported by wrapping the body in a Hash with the `:params`
key. Not ideal, but works for now. If we set a hash directly it's
ignored, because we fetch and filter the params by first wrapping it in
a `Appsignal::Transaction::GenericRequest` instance. We can't set a
params value that is not a Hash directly.

We recommend that users use a
`body_parser` for their workers that parses the body in an acceptable
Hash format. The recommended parser is the JSON parser.

```
class MyWorker
  include Shoryuken::Worker

  shoryuken_options body_parser: JSON

  def perform(sqs_msg, body)
    puts body
  end
end
```