Add background job integrations argument filtering
Previously we truncated long values in the background jobs to prevent
apps from sending large payloads with data that's unused and doesn't
provide more context. This logic was added to the parameter filtering in
the commit before this.

This change utilizes the parameter filtering for the background job
integrations as well. Which adds support for filtering arguments to
background job libraries and also truncates large argument values.