Fix raise and throws in instrument helper
By using a "soft return", by saving the return value in a variable, it
can happen that an event is not finished when the block that's given to
an instrument method raises and error or uses `throw` to throw a signal.
Using `ensure` we _ensure_ that the event is finished even when that
happens.