Shoryuken instrumentation body arg can be a string (#266)
* The body arg can be a string

If Shoryuken option for `body_parser` is `:text` or `nil` (or just not passed), then `body` will be a `String` and it will crash at `body.reject {...}`

* Add specs for non-hash/array body arg
Also fix issue later down when value is fetched from body via key (should not crash but also doesn't make sense)
-----------------------------------------------------------
On branch master - Thu 16 Mar 2017 17:04:22 PDT by matrinox <geofflee25@gmail.com>

* Change exclude_keys to a class const
Check body is a hash instead of specific method
-----------------------------------------------------------
On branch master - Mon  3 Apr 2017 14:32:51 PDT by matrinox <geofflee25@gmail.com>

* Update Shoryuken integration

Fix the Shoryuken integration. Looks like the initial version was based
on our Sidekiq integration, but it does not work like that.

Updated to work against version 1 through 3 with Shoryuken and this was
the behavior I could determine from the library.

The original request that brought the problem to our attention was to
support string (or any other type than a Hash) as a body for the worker.
This is now supported by wrapping the body in a Hash with the `:params`
key. Not ideal, but works for now. If we set a hash directly it's
ignored, because we fetch and filter the params by first wrapping it in
a `Appsignal::Transaction::GenericRequest` instance. We can't set a
params value that is not a Hash directly.

We recommend that users use a
`body_parser` for their workers that parses the body in an acceptable
Hash format. The recommended parser is the JSON parser.

```
class MyWorker
  include Shoryuken::Worker

  shoryuken_options body_parser: JSON

  def perform(sqs_msg, body)
    puts body
  end
end
```

* Use format_args to prevent large XML files from being formed
-----------------------------------------------------------
On branch master - Mon 17 Apr 2017 10:57:08 PDT by matrinox <geofflee25@gmail.com>

* Upgrade from format_args to Appsignal::Utils::ParamsSanitizer
-----------------------------------------------------------
On branch master - Tue 18 Apr 2017 10:57:31 PDT by matrinox <geofflee25@gmail.com>

* Truncate and filter Shoryuken arguments

Previously we truncated long values in the background jobs to prevent
apps from sending large payloads with data that's unused and doesn't
provide more context.

This change utilizes the parameter filtering for the Shoryuken
background job integration as well. Which adds support for filtering
arguments and also truncates large argument values.

* Update exception testing for Shoryuken

* Remove lengthy parameter values testing