Fix queue time on DelayedJob integration (#297)
Would use `created_at`, which is the time the job was created. This is
not necessarily when it was meant for the queue. Yes, it is enqueued in
the sense that it is registered in the job queue list, but it's only
meant to actually be enqueued in the DelayedJob worker at the time
specified in `run_at`.

This change uses the run_at as the queue start time so that jobs that
are scheduled to run a long time from now don't add the time from the
moment the job was created until it was actually run as the queue time.