Fix raise and throw in ActiveSupport Notifications
By using a "soft return", by saving the return value in a variable, it
can happen that an event is not finished when the block that's given to
an instrument method raises and error or uses throw to throw a signal.
Using ensure we ensure that the event is finished even when that
happens.

Follow up from #292.