Add ignore_namespaces option
Closes #272.

This feature was added to the AppSignal extension in PR
https://github.com/appsignal/appsignal-agent/pull/255, this PR adds the
config option to the Ruby gem.

Configure `ignore_namespaces` to ignore transactions (requests and jobs)
from certain namespaces configured as described in our docs:
http://docs.appsignal.com/application/namespaces.html

This allows users to ignore entire namespaces, such as administration
panels, without having to ignore all the controllers in that namespace
with `ignore_actions`.