Add support for standalone ActionCable servers
Standalone ActionCable servers don't have an ActionDispatch request_id
in the request env. Instead, we create our own and add it to the request
under a private key. This key is then used to save the request id, also
for non-standalone servers, and reused throughout the duration of the
websocket. This allows us to track all messages on a websocket under the
same request_id. This could make searching for all samples on one
websocket connection a lot easier.