Fix hanging installer & diagnose on Windows
Write output immediately to STDOUT on Microsoft Windows.
Without `$stdout.sync = true` the output would be saved until the end of
the program and then flushed to STDOUT.

When we print the output immediately we don't get the illusion of a
hanging installer and diagnose task when we ask for user input. Because
the user input prompt hasn't been flushed to STDOUT the user doesn't see
what's expected of them to do. This change fixes that.