Fix Capistrano config overrides
Previously any config given to the Capistrano deploy definition was
given as the initial config of Appsignal::Config. This initial config is
then overridden by any next step of the config: config file and env
variables, which made it kind of useless. Instead, override the config
with the Capistrano config after all the other config is loaded.