Track Exception-level exceptions in integrations
This allows us to track more Exceptions in Ruby without breaking the
normal Exception handling in Ruby. The main thing to keep in mind while
rescuing Exception is that you re-raise the Exception in case it's an
`SystemExit` or `SignalException`, otherwise we might break the normal
shutdown behavior of Ruby processes.

*Note:* Webmachine is not updated yet as it breaks something by raising
an Exception-level Exception.