Fix DelayedJob ActiveJob integration
I think there was a wrong assumption about how ActiveJob jobs are
wrapped in ActiveJob. The object being passed to
`DelayedJobPlugin.invoke` is not the
`ActiveJob::QueueAdapters::DelayedJobAdapter::JobWrapper`, instead the
payload object is wrapped in the `JobWrapper`. The DelayedJob "backend"
stays the same and so does the object being passed to
`DelayedJobPlugin.invoke`, just the payload on the job differs.