Record Sidekiq delayed extension actions properly
Now it no longer reports class method (with a `.` separator) (used by
the Sidekiq delayed extension) as instance methods (with a `#`
separator).

`DelayedExtensionWorker#slow` becomes `DelayedExtensionWorker.slow`.

"Normal" jobs with a dedicated class and using the `perform` instance
method, are still reported as `MyWorker#perform`.

Updated the specs to use the new-ish `transaction#to_h` method to avoid
having to stub so much in the specs. Noticed the returned data is
somewhat incomplete for data on the root object of the transaction, such
as `queue_start` and `queue_time`. Separate issues have been created for
those.

This still doesn't fix the fact that we don't actually test against
Sidekiq itself and our tests assume a certain data format used by
Sidekiq, which may change over time.