Add tags and namespace options to listen_for_error (#357)
`Appsignal.listen_for_error` is the only helper that doesn't support
custom tags and a custom namespace.
This change adds support for the tags and namespace and passes the
values along to `Appsignal.send_error`.

Did not change this spec to use `Appsignal::Transaction#to_h`, because
the change would be a lot more involved as we can't fetch the current
transaction from the current thread. To record a transaction we'd need
to be able to capture the transaction as it's being created.