Support Sidekiq delay extension
Import logic from the sidekiq/api library for display_class and
display_args. This adds support for the Sidekiq delay extension and
better supports most ActiveJob classes, along with ActionMailer classes.

This is the second attempt at fixing this. The first attempt directly
used classes from sidekiq/api, Sidekiq::Job to get the class name and
arguments.

Downside to this was the dependency on a internal Sidekiq class and
loss of backwards compatibility for certain older versions.