Support Sidekiq encrypted values
In Sidekiq enterprise secret bag values are supported as the last
argument for a job. When this is the case the job is marked as encrypted
with the `"encrypt"` key set to `true`.

None of the other values are touched, only the last one is encrypted and
replaced by AppSignal with the `"[encrypted data]"` string, like how the
Sidekiq web UI displays it.