Write install platform on install
This allows us to make sure that the diagnose command reports the
platform for which it was last installed, rather than the platform it
detects during the `appsignal diagnose` run. This may differ as the
`APPSIGNAL_BUILD_FOR_MUSL` environment variable may not be present
during the diagnose run, but it is during the installation.