Store agent architecture rather than platform
The agent architecture is the combination of CPU type and build
platform. This change stores the architecture to the
`ext/appsignal.architecture` file and includes this in the diagnose
report.

Why save the architecture? As gives a more complete picture of what was
installed? Did it accidentally install the 32-bit version or the 64-bit
version? With this change we'll know and we'll know exactly which
package it used from the `ext/agent.yml` file, which allows us to debug
the install even better.