Remove Set usage (#373)
In PR #362 the Sidekiq hook `job_keys` method was moved to the
`JOB_KEYS` constant. As its value is a `Set` instance the `set` library
is required to be loaded beforehand. This was loaded by Sidekiq in
https://github.com/mperham/sidekiq/blob/58d5ff1e6d96df4a0b3a53d74a8aa26cf08dc467/lib/sidekiq/manager.rb#L7
which is loaded before the AppSignal hooks is called.

The `JOB_KEYS` constant doesn't have to be of type `Set`, we don't use
it any different from a normal array. This change removes the set type
from the `JOB_KEYS` constant.

This fixes the problem of AppSignal CLI crashing on start:

```
bin/appsignal diagnose
/appsignal/gem/lib/appsignal/hooks/sidekiq.rb:25:in
  `<class:SidekiqPlugin>': uninitialized constant
  Appsignal::Hooks::SidekiqPlugin::Set (NameError)
```