Add jRuby support
Use FFI to link to functions from the Rust extension. Add a small Ruby
wrapper for the Extension, Transaction and Data types so the interface
stays the same between the C-extension and FFI. By using constant
reassignment in `lib/appsignal/extension.rb` we can keep the docs
intact for both implementations.

Add a new build to work with empty strings for configuration options,
they are now considered unset config options. Was necessary to work
around something that appears to be a jRuby bug when unsetting keys
on the `ENV` hash.