Fix jRuby publish by making a custom gemspec
A problem with publishing the jRuby version of the gem is that you will
need to have jRuby installed locally in order for it to pick up the
gemspec file. This is because it's only specified as `gem.platform =
"java"` if the Ruby version is jRuby.

When you're running a non-jRuby Ruby version it will not find the gem
version for jRuby when publishing the gem. This causes the `rake
publish` task to error and exit early, only shipping the normal Ruby gem
version.

Instead, configure the gemspec in the Rake tasks dynamically and
configure a `Gem::PackageTask` for that gemspec. Then, when calling
those special tasks it will build the correct version, allowing us to
ship it without needing to switch to jRuby in the Rake task itself
(which would require knowledge of which Ruby version manager is
installed locally).