Fix jRuby specs with new agent build
We always need to unset the diagnose variable. For normal Ruby this is
unset in the diagnose task itself, but the jRuby bug requires us to
unset it using the method below as well. It's not present in the ENV
keys list because it's already cleared in Ruby itself in the diagnose
task, so add it manually to the list of to-be cleaned up keys.