Add filter_session_data config option
Create a separate option to filter session data and ensure we don't
overuse the `filter_parameters` option for multiple sources of data.

This is an addition on PR #399.

Also refactor the tests a bit to ensure we also test the parameter
filtering ActionDispatch session objects.

Actually use the ParamSanitizer rather than mocking the result.