Fall back on job arguments for ActiveJob Sidekiq
When libraries forget to set the `wrapped` value in the job payload,
fall back on the first argument to extract the job class name,
arguments and other metadata.

To do so we need to ensure the value from the argument can only be
fetched from the `job_class` key and is a String object.