Remove default hostname (#404)
Instead we'll base determine this default as a fallback in the
extension. The integrations don't need to know about the default
hostname, also reduces the logic in the integrations.

The integrations will only set the hostname config option in the
environment `_APPSIGNAL_HOSTNAME` when it has been configured in the
app. If not, it will set an empty string, which is interpreted as an
empty value by the extension.