Log unset request_headers option warning to STDOUT
This makes it better visible for users that the behavior will change and
that they need to set the option themselves from now on.

This change required some updates on the test suite. It now silences
this warning by setting the request_headers option for tests that would
print the warning in the test output. One test now has a silencer around
it, because setting the value as an env var would distract from the
test.

The `silence` helper is now updated to raise errors in the output as an
error that would fail the test suite.