Refactor Transaction#set_* methods specs (#420)
Test the actual data stored on the transaction with `Transaction#to_h`
and on the Transaction object.

Also found a broken assertion on `Appsignal::Extension` calls. Needs to
be performed on an instance of the extension class, not the class
itself.

This won't need to be fixed anymore as we check the value on the
transaction in the extension.

Improper calls, as we do in the test suite, would cause an error to be
raised if not handled correctly. For example:

```
Failure/Error: @ext.set_action(nil)

TypeError:
  wrong argument type nil (expected String)
```