Add error message for unsupported architectures
In case someone uses an architecture that we don't support we want to
let them know instead of showing an exception.

For example, when installing AppSignal for JRuby with
`APPSIGNAL_BUILD_FOR_MUSL=1` set it would fail the installation with a
non-descriptive error and backtrace. While we do not support the JRuby
and Musl combination, at least show a more descriptive message for the
user (and us) about what went wrong.

Fixes #425