Add diagnose command --[no-]send-report option
This option allows users to (not) send the report without being
prompted. Useful for our support team to send this all in one command to
the user, so users don't have to confirm anything.

```
appsignal diagnose --send-report
appsignal diagnose --no-send-report
```

Also useful for non-interactive environments in which the command is
run.