Diagnose command also logs to logger
Start the logger by calling `start_logger`. Log an info message to the
logger not only for the test, but also as a marker for the actual log
contents so we know where what comes from in real logs.

Update the tests to assert the log file has been created, rather that
"Exists?: false". The tests now create the log file, so it needs to
check a different type of output from the diagnose command, because
"Exists?: false" is only printed if the file doesn't exist.