Silence logger output in tests
Now that the diagnose command starts the AppSignal logger it prints some
errors to STDOUT when no writable log file can be found. The config
prints a valid error, but it's part of the scenario we are testing.

Expand the `silence` helper to allow certain errors and wrap the
`Config.new` calls with the `silence` helper. This way we silence the
output in the tests while still triggering an error if a not allowed
error is logged.