Reduce number of TravisCI builds
Speed up the TravisCI build, by running less jobs.

For Ruby versions that are End Of Life, only run the `no_dependencies`
Gemfile. Except for Ruby 2.0, the oldest version we test against, test
as much as possible there so we know if all our  changes work on the
oldest version.

Note that Rubocop already checks if we use valid Ruby 1.9 syntax. So
syntax errors isn't what's being tested, just if it works at all on Ruby
2.0.