Move paths and files diagnose to separate classes
This cleans up the main diagnose class a bit so we can add other classes
for sub sections more easily. The main diagnose class will be primarily
report construction from separate classes and printing the output to the
terminal.

The new paths setup will merge directories and files and as a report
section. It will read the file contents if the path is about a file.
This means the `appsignal.log` file contents will now be send to
AppSignal servers.

Directories and files now have the exact same data structure which makes
it easier to print them to the terminal and parse them in the AppSignal
admin.

Moved the `library.package_install_path` path to
`paths.package_install_path` so we also perform all the directory checks
on the gem install path.

New additions to the data structure:
- `type` - type of path ("directory" / "file")
- `mode` - file permissions (in addition to only sending `writable`)
- `ownership`
  - `gid` - group id of the path
  - `group` - group name of the path

Refactored some specs to not test the `Appsignal::Config#log_file_path`
logic again in the `diagnose_spec.rb`.