Track config options sources
For diagnose purposes it's useful to know from what source the config
option is set. For example, this helps debugging file config being
overwritten by the system environment.

Change the report format to include all the sources as well as the final
config used by the gem.

Update how the config is loaded in Config#initialize to work the same
way for every source. Change `#merge` to directly update the
`config_hash`, no need to pass it along if we already which config we're
updating.

Looks like there were no specs for the config section in the transmitted
report. Added that now with the new structure.