Link back to AppSignal diagnose report page (#445)
When a user sends a report our way we link back to the AppSignal
diagnose report view page. Initially users will be able to "claim" their
report, so we know who to contact.

Step 2 is making our report validator public so users can see any
validation failures and start debugging themselves, but already keep the
message as such for now so we don't have to ship a new release with an
updated message.