Build custom .travis.yml
To fix the builds on TravisCI we need a bit more flexibility to specify
the required RubyGems and Bundler versions.

For Padrino builds we need a Bundler version lower than 2.0 for example.
Some ActiveSupport (Rails) versions require Bundler < 2.0 as well.

This new matrix also decrease the number of jobs by choosing "gemsets"
per Ruby version. We don't have to test every Ruby version against every
gem combination. That's more the responsibility of the gems themselves.
And we do not have to support a combination if the gem itself doesn't.

By running `rake travis:generate` a new `.travis.yml` file is generated
we use to build against.

To ensure the build is using the latest intended build matrix, validate
the build matrix as the first job to see if the `.travis.yml` generated
differs from the committed version.